Fix float rounding issues (#736)
* Fix rounding bugs

* Mypy fixes

* Undo some changes

* Undo some changes

* mypy fixes

* Fix tests

* Increase size of ETCI2021 fake data

* Undo model changes

* line length fix

+++ torchgeo/datasets/spacenet.py
@@ -703,7 +703,7 @@ class SpaceNet3(SpaceNet):
         speed_arr_bin = np.arange(min_speed_bin, max_speed_bin + 1)
         bin_size_mph = 10.0
         speed_cls_arr: "np.typing.NDArray[np.int_]" = np.array(
-            [int(math.ceil(s / bin_size_mph)) for s in speed_arr_bin]
+            [math.ceil(s / bin_size_mph) for s in speed_arr_bin]
         )
 
         try:
